Updates options to use string collection instead of CSV
