+++ Italics.sln
@@ -8,17 +8,23 @@ EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|arm64 = Debug|arm64
 		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|arm64 = Release|arm64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Debug|arm64.Build.0 = Debug|Any CPU
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Debug|x86.ActiveCfg = Debug|x86
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Debug|x86.Build.0 = Debug|x86
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Release|arm64.ActiveCfg = Release|Any CPU
+		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Release|arm64.Build.0 = Release|Any CPU
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Release|x86.ActiveCfg = Release|x86
 		{7A2F8C00-6D86-4041-AF9C-AA00D44C8B61}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
